feat: Add LeafNode class
This commit introduces the LeafNode class, which represents a leaf node in the HTML tree.

A leaf node is a node that cannot have any children. It inherits from the HTMLNode class and overrides the to_html method to generate the HTML representation of the node.

The LeafNode class is used to represent HTML elements that do not have any content, such as <img> and <br>, as well as elements that contain text, such as <p> and <h1>.